api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
transformers
Pillow
numpy
opencv-python
